# test_output/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ilma API",
    "description": "School management system API with end-to-end encryption for messaging. Currently under development.",
    "contact": {
      "name": "API Support / Developer",
      "email": "jero.lampila@gmail.com"
    },
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/me": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get_me",
        "responses": {
          "200": {
            "description": "Current authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/user/public-key/{username}": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get_user_public_key_by_username",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User's public key retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicKeyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/user/recovery-key/{username}": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get_recovery_key",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recovery key retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryKeyResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — admin access only"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/user/set-recovery-key": {
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "set_recovery_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRecoveryKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Recovery key updated"
          },
          "400": {
            "description": "Invalid data or unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users": {
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/users/{id}/public_key": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get_user_public_key",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User's public key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicKeyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "username",
          "email",
          "password",
          "role",
          "public_key"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "public_key": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "PublicKeyResponse": {
        "type": "object",
        "required": [
          "public_key"
        ],
        "properties": {
          "public_key": {
            "type": "string"
          }
        }
      },
...
      "RecoveryKeyResponse": {
        "type": "object",
        "required": [
          "recovery_key"
        ],
        "properties": {
          "recovery_key": {
            "type": "string"
          }
        }
      },
...
      "SetRecoveryKeyRequest": {
        "type": "object",
        "required": [
          "username",
          "recovery_key"
        ],
        "properties": {
          "recovery_key": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "User": {
        "type": "object",
        "required": [
          "id",
          "username",
          "email",
          "role",
          "is_superuser",
          "public_key",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_superuser": {
            "type": "boolean"
          },
          "public_key": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "student",
          "teacher",
          "principal"
        ]
      }
    }
  },
...
}